+++ tests/testthat/test-bind_rows-lazy.R
@@ -95,4 +95,31 @@ test_that("arg .id works", {
   )
 })
 
+test_that("error if not all elements don't have the same class", {
+  p1 <- pl$LazyFrame(
+    x = sample(letters, 20),
+    y = sample(1:100, 20)
+  )
+  p2 <- data.frame(
+    x = sample(letters, 20),
+    y = sample(1:100, 20)
+  )
+
+  expect_snapshot_lazy(bind_rows_polars(p1, p2), error = TRUE)
+})
+
+test_that("elements must be either all DataFrames or all LazyFrames", {
+  skip_if(Sys.getenv("TIDYPOLARS_TEST") == "TRUE")
+  p1 <- pl$LazyFrame(
+    x = sample(letters, 20),
+    y = sample(1:100, 20)
+  )
+  p2 <- pl$LazyFrame(
+    x = sample(letters, 20),
+    y = sample(1:100, 20)
+  )
+
+  expect_snapshot_lazy(bind_rows_polars(p1, p2), error = TRUE)
+})
+
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)