+++ tests/testthat/test-bind_rows.R
@@ -90,3 +90,30 @@ test_that("arg .id works", {
     as.character(rep(1:2, each = 20))
   )
 })
+
+test_that("error if not all elements don't have the same class", {
+  p1 <- pl$DataFrame(
+    x = sample(letters, 20),
+    y = sample(1:100, 20)
+  )
+  p2 <- data.frame(
+    x = sample(letters, 20),
+    y = sample(1:100, 20)
+  )
+
+  expect_snapshot(bind_rows_polars(p1, p2), error = TRUE)
+})
+
+test_that("elements must be either all DataFrames or all LazyFrames", {
+  skip_if(Sys.getenv("TIDYPOLARS_TEST") == "TRUE")
+  p1 <- pl$DataFrame(
+    x = sample(letters, 20),
+    y = sample(1:100, 20)
+  )
+  p2 <- pl$LazyFrame(
+    x = sample(letters, 20),
+    y = sample(1:100, 20)
+  )
+
+  expect_snapshot(bind_rows_polars(p1, p2), error = TRUE)
+})